backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.0
pydantic-settings==2.5.0
openai==1.50.0
python-multipart==0.0.9
httpx==0.27.0
chromadb==0.5.0
gitpython==3.1.43
tiktoken==0.7.0
cryptography==43.0.0
